Move extra files into an exploration file, which is just files used for data exploration and results.
